stock-analysis-agent/pyproject.toml
[project]
name = "stock-analysis-agent"
version = "0.1.0"
description = "AI-driven multi-agent system for comprehensive stock analysis"
authors = [
    {name = "JayVisaria9",email = "jay.visaria@pfizer.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.4.2,<2.0.0)",
    "google-cloud-aiplatform (>=1.98.0,<2.0.0)",
    "google-genai (>=1.21.1,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
